Magento sync: civettaai attribute + operational/observation mode + 2h cron
- Fix: attributo Magento corretto è `civettaai` (non `api_civettaai`,
  inesistente sui Magento target). Prima i PUT venivano silenziosamente
  ignorati. Option_id letti runtime via /rest/V1/products/attributes,
  cachati 24h per tenant.
- Modello mode per tenant in tenant_configs.xhumanpro_magento_mode:
  - observation: xHumanPro calcola e mostra cosa farebbe (preview) ma
    non scrive su Magento. Farmabooster controlla via attributo `civetta`.
  - operational: xHumanPro scrive `civettaai`, prevale su Farmabooster.
- Delta fetch pre-PUT: legge stato corrente di tutti i candidati, fa
  PUT solo dove diverge dal target. Su Papa: 378 PUT effettivi su 10.296
  candidati (-97%), tempo 5 min vs 40 min. civettaai=null trattato come
  "FB controlla" → forzato sempre al target esplicito per priorità.
- try/finally in execute() garantisce persistenza di feed_actions.applied
  + feed_dispatch_log anche in caso di crash a metà run.
- magentoSyncCron: ogni 2h al :30, sequenziale per tenant, solo
  operational. Wiring in server.js.
- Fix collaterali: zombieCron 06:00→03:00 UTC (dato TP completo entro
  05:00 italia), logger auto-extract tenant/source dal payload,
  auth+tenant middleware su /logs e /cross-channel, feed_actions ON
  CONFLICT DO NOTHING per duplicati cross-categoria, civetta_sync
  parallelizzato (pageSize 300→1000, batch 3).

+++ backend/services/abTestService.js
@@ -248,7 +248,7 @@ async function evaluateOne(test) {
   const orderRatio = expectedOrders > 0 ? m.orders / expectedOrders : 0;
 
   await pool.query(
-    `UPDATE ab_tests SET test_clicks=$1, test_orders=$2, test_revenue=$3,
+    `UPDATE ab_tests SET test_clicks=$1::int, test_orders=$2::int, test_revenue=$3::numeric,
        test_cpa = CASE WHEN $2::int > 0 THEN $1::numeric/$2::numeric ELSE NULL END
      WHERE id=$4`,
     [m.clicks, m.orders, m.revenue, test.id]);
